// __generated__/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
};

/** A Field Group registered by ACF */
export type AcfFieldGroup = {
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
};
// ...
export type Connection = {
  /** A list of edges (relational context) between connected nodes */
  edges: Array<Edge>;
  /** A list of connected nodes */
  nodes: Array<Node>;
  /** Information about pagination in a connection. */
  pageInfo: PageInfo;
};
// ...
export type PageConnection = {
  /** A list of edges (relational context) between RootQuery and connected page Nodes */
  edges: Array<PageConnectionEdge>;
  /** A list of connected page Nodes */
  nodes: Array<Page>;
  /** Information about pagination in a connection. */
  pageInfo: PageConnectionPageInfo;
};
// ...
export type Page_Veterinary = AcfFieldGroup & {
  __typename?: 'Page_Veterinary';
  caution?: Maybe<Scalars['String']['output']>;
  description?: Maybe<Scalars['String']['output']>;
  descriptionAtFooter?: Maybe<Scalars['String']['output']>;
  faqs?: Maybe<Page_Veterinary_Faqs>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  procedure?: Maybe<Scalars['String']['output']>;
  reasons?: Maybe<Page_Veterinary_Reasons>;
  schedulingForDeworming?: Maybe<Scalars['String']['output']>;
  whatWeUse?: Maybe<Scalars['String']['output']>;
  whyPetShouldDeworming?: Maybe<Page_Veterinary_WhyPetShouldDeworming>;
};

/** Field Group */
export type Page_Veterinary_Faqs = AcfFieldGroup & {
  __typename?: 'Page_Veterinary_Faqs';
  faq1?: Maybe<Page_Veterinary_Faqs_Faq1>;
  faq2?: Maybe<Page_Veterinary_Faqs_Faq2>;
  faq3?: Maybe<Page_Veterinary_Faqs_Faq3>;
  faq4?: Maybe<Page_Veterinary_Faqs_Faq4>;
  faq5?: Maybe<Page_Veterinary_Faqs_Faq5>;
  faq6?: Maybe<Page_Veterinary_Faqs_Faq6>;
  faq7?: Maybe<Page_Veterinary_Faqs_Faq7>;
  faq8?: Maybe<Page_Veterinary_Faqs_Faq8>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
};

/** Field Group */
export type Page_Veterinary_Faqs_Faq1 = AcfFieldGroup & {
  __typename?: 'Page_Veterinary_Faqs_Faq1';
  answer?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  question?: Maybe<Scalars['String']['output']>;
};

/** Field Group */
export type Page_Veterinary_Faqs_Faq2 = AcfFieldGroup & {
  __typename?: 'Page_Veterinary_Faqs_Faq2';
  answer?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  question?: Maybe<Scalars['String']['output']>;
};

/** Field Group */
export type Page_Veterinary_Faqs_Faq3 = AcfFieldGroup & {
  __typename?: 'Page_Veterinary_Faqs_Faq3';
  answer?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  question?: Maybe<Scalars['String']['output']>;
};

/** Field Group */
export type Page_Veterinary_Faqs_Faq4 = AcfFieldGroup & {
  __typename?: 'Page_Veterinary_Faqs_Faq4';
  answer?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  question?: Maybe<Scalars['String']['output']>;
};

/** Field Group */
export type Page_Veterinary_Faqs_Faq5 = AcfFieldGroup & {
  __typename?: 'Page_Veterinary_Faqs_Faq5';
  answer?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  question?: Maybe<Scalars['String']['output']>;
};

/** Field Group */
export type Page_Veterinary_Faqs_Faq6 = AcfFieldGroup & {
  __typename?: 'Page_Veterinary_Faqs_Faq6';
  answer?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  question?: Maybe<Scalars['String']['output']>;
};

/** Field Group */
export type Page_Veterinary_Faqs_Faq7 = AcfFieldGroup & {
  __typename?: 'Page_Veterinary_Faqs_Faq7';
  answer?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  question?: Maybe<Scalars['String']['output']>;
};

/** Field Group */
export type Page_Veterinary_Faqs_Faq8 = AcfFieldGroup & {
  __typename?: 'Page_Veterinary_Faqs_Faq8';
  answer?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  question?: Maybe<Scalars['String']['output']>;
};

/** Field Group */
export type Page_Veterinary_Reasons = AcfFieldGroup & {
  __typename?: 'Page_Veterinary_Reasons';
  description?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  reason1?: Maybe<Page_Veterinary_Reasons_Reason1>;
  reason2?: Maybe<Page_Veterinary_Reasons_Reason2>;
  reason3?: Maybe<Page_Veterinary_Reasons_Reason3>;
  reason4?: Maybe<Page_Veterinary_Reasons_Reason4>;
};

/** Field Group */
export type Page_Veterinary_Reasons_Reason1 = AcfFieldGroup & {
  __typename?: 'Page_Veterinary_Reasons_Reason1';
  description?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  title?: Maybe<Scalars['String']['output']>;
};

/** Field Group */
export type Page_Veterinary_Reasons_Reason2 = AcfFieldGroup & {
  __typename?: 'Page_Veterinary_Reasons_Reason2';
  description?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  title?: Maybe<Scalars['String']['output']>;
};

/** Field Group */
export type Page_Veterinary_Reasons_Reason3 = AcfFieldGroup & {
  __typename?: 'Page_Veterinary_Reasons_Reason3';
  description?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  title?: Maybe<Scalars['String']['output']>;
};

/** Field Group */
export type Page_Veterinary_Reasons_Reason4 = AcfFieldGroup & {
  __typename?: 'Page_Veterinary_Reasons_Reason4';
  description?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  title?: Maybe<Scalars['String']['output']>;
};

/** Field Group */
export type Page_Veterinary_WhyPetShouldDeworming = AcfFieldGroup & {
  __typename?: 'Page_Veterinary_WhyPetShouldDeworming';
  description?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  reason1?: Maybe<Page_Veterinary_WhyPetShouldDeworming_Reason1>;
  reason2?: Maybe<Page_Veterinary_WhyPetShouldDeworming_Reason2>;
  reason3?: Maybe<Page_Veterinary_WhyPetShouldDeworming_Reason3>;
  reason4?: Maybe<Page_Veterinary_WhyPetShouldDeworming_Reason4>;
};

/** Field Group */
export type Page_Veterinary_WhyPetShouldDeworming_Reason1 = AcfFieldGroup & {
  __typename?: 'Page_Veterinary_WhyPetShouldDeworming_Reason1';
  answer?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  question?: Maybe<Scalars['String']['output']>;
};

/** Field Group */
export type Page_Veterinary_WhyPetShouldDeworming_Reason2 = AcfFieldGroup & {
  __typename?: 'Page_Veterinary_WhyPetShouldDeworming_Reason2';
  answer?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  question?: Maybe<Scalars['String']['output']>;
};

/** Field Group */
export type Page_Veterinary_WhyPetShouldDeworming_Reason3 = AcfFieldGroup & {
  __typename?: 'Page_Veterinary_WhyPetShouldDeworming_Reason3';
  answer?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  question?: Maybe<Scalars['String']['output']>;
};

/** Field Group */
export type Page_Veterinary_WhyPetShouldDeworming_Reason4 = AcfFieldGroup & {
  __typename?: 'Page_Veterinary_WhyPetShouldDeworming_Reason4';
  answer?: Maybe<Scalars['String']['output']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  question?: Maybe<Scalars['String']['output']>;
};
// ...
export type RootQueryToPageConnection = Connection & PageConnection & {
  __typename?: 'RootQueryToPageConnection';
  /** Edges for the RootQueryToPageConnection connection */
  edges: Array<RootQueryToPageConnectionEdge>;
  /** The nodes of the connection, without the edges */
  nodes: Array<Page>;
  /** Information about pagination in a connection. */
  pageInfo: RootQueryToPageConnectionPageInfo;
};
// ...
export type Veterianry_FragmentFragment = { __typename?: 'Page', veterinary?: { __typename?: 'Page_Veterinary', description?: string | null, whatWeUse?: string | null, caution?: string | null, procedure?: string | null, schedulingForDeworming?: string | null, faqs?: { __typename?: 'Page_Veterinary_Faqs', faq1?: { __typename?: 'Page_Veterinary_Faqs_Faq1', question?: string | null, answer?: string | null } | null, faq2?: { __typename?: 'Page_Veterinary_Faqs_Faq2', question?: string | null, answer?: string | null } | null, faq3?: { __typename?: 'Page_Veterinary_Faqs_Faq3', question?: string | null, answer?: string | null } | null, faq4?: { __typename?: 'Page_Veterinary_Faqs_Faq4', question?: string | null, answer?: string | null } | null, faq5?: { __typename?: 'Page_Veterinary_Faqs_Faq5', question?: string | null, answer?: string | null } | null, faq6?: { __typename?: 'Page_Veterinary_Faqs_Faq6', question?: string | null, answer?: string | null } | null, faq7?: { __typename?: 'Page_Veterinary_Faqs_Faq7', question?: string | null, answer?: string | null } | null, faq8?: { __typename?: 'Page_Veterinary_Faqs_Faq8', question?: string | null, answer?: string | null } | null } | null, reasons?: { __typename?: 'Page_Veterinary_Reasons', description?: string | null, reason1?: { __typename?: 'Page_Veterinary_Reasons_Reason1', title?: string | null, description?: string | null } | null, reason2?: { __typename?: 'Page_Veterinary_Reasons_Reason2', title?: string | null, description?: string | null } | null, reason3?: { __typename?: 'Page_Veterinary_Reasons_Reason3', title?: string | null, description?: string | null } | null, reason4?: { __typename?: 'Page_Veterinary_Reasons_Reason4', title?: string | null, description?: string | null } | null } | null, whyPetShouldDeworming?: { __typename?: 'Page_Veterinary_WhyPetShouldDeworming', description?: string | null, reason1?: { __typename?: 'Page_Veterinary_WhyPetShouldDeworming_Reason1', question?: string | null, answer?: string | null } | null, reason2?: { __typename?: 'Page_Veterinary_WhyPetShouldDeworming_Reason2', question?: string | null, answer?: string | null } | null, reason3?: { __typename?: 'Page_Veterinary_WhyPetShouldDeworming_Reason3', question?: string | null, answer?: string | null } | null, reason4?: { __typename?: 'Page_Veterinary_WhyPetShouldDeworming_Reason4', question?: string | null, answer?: string | null } | null } | null } | null } & { ' $fragmentName'?: 'Veterianry_FragmentFragment' };
// ...
export type Get_VeterinaryQueryVariables = Exact<{
  name: Scalars['String']['input'];
}>;


export type Get_VeterinaryQuery = { __typename?: 'RootQuery', pages?: { __typename?: 'RootQueryToPageConnection', nodes: Array<(
      { __typename?: 'Page' }
      & { ' $fragmentRefs'?: { 'Fragment_Page_TypeFragment': Fragment_Page_TypeFragment;'Veterianry_FragmentFragment': Veterianry_FragmentFragment } }
    )> } | null };
// ...
export type Fragment_Page_TypeFragment = { __typename?: 'Page', featuredImage?: { __typename?: 'NodeWithFeaturedImageToMediaItemConnectionEdge', node: (
      { __typename?: 'MediaItem' }
      & { ' $fragmentRefs'?: { 'Fragment_Media_ItemFragment': Fragment_Media_ItemFragment } }
    ) } | null, seo?: { __typename?: 'PostTypeSEO', title?: string | null, metaDesc?: string | null } | null } & { ' $fragmentName'?: 'Fragment_Page_TypeFragment' };